+++ ReleaseNotes.md
@@ -4,6 +4,7 @@
 - [Permanent redirects are now in the legacy pages](https://trello.com/c/vJnpgZol/211-redirect-from-old-ad-view-and-search-result-pages)
 - [Contact phone and email replace the value and type now](https://trello.com/c/NjuH39xM/213-replace-the-contact-value-and-type-with-phone-and-email)
 - [Ad view labels for phone and email now have a cool label for mobile phone so it can dial automatically](https://trello.com/c/XxlQE1id/208-contact-details-on-ad-view-page-could-use-direct-email-and-phone-links)
+- [The support team email template is updated]
 
 **Bug Fixes**